Add EnvAgentInterfaceCard and refactor agent overview sections
